feat(httploader): add allowed source regexp option
This commit introduces a new option for the httploader that allows
to specify a regexp that will be used to filter the sources that
are allowed to be loaded.

It is useful for whitelisting paths that can't be covered by the
host based filtering, for example when wanting to load from multiple
GCS buckets by URL, and GCS bucket names are part of the path in the
URL unlike S3 bucket names that are part of the host name.